LL-1778 Fix show release notes after update
